+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Renamed `delete_results_dataset` param to `delete_results_datasets` in `Beaker.experiment.delete()`.
+
+### Fixed
+
+- Fixed bug with `Beaker.experiment.delete()` that would lead to a `ValueError` being raised when the experiment to delete has multiple tasks.
+
 ## [v0.14.1](https://github.com/allenai/beaker-py/releases/tag/v0.14.1) - 2022-05-05
 
 ### Changed